+++ android/src/main/kotlin/ru/rustore/flutter_rustore_push/FlutterRustorePushService.kt
@@ -1,55 +1,76 @@
 package ru.rustore.flutter_rustore_push
 
-import ru.rustore.flutter_rustore_billing.utils.Log
-import ru.rustore.flutter_rustore_push.pigeons.Rustore
+import android.os.Handler
+import android.os.Looper
+import ru.rustore.flutter_rustore_push.pigeons.Message
+import ru.rustore.flutter_rustore_push.pigeons.Notification
+import ru.rustore.flutter_rustore_push.pigeons.RuStorePushCallbacks
+import ru.rustore.flutter_rustore_push.utils.Log
 import ru.rustore.sdk.pushclient.messaging.exception.RuStorePushClientException
 import ru.rustore.sdk.pushclient.messaging.model.RemoteMessage
 import ru.rustore.sdk.pushclient.messaging.service.RuStoreMessagingService
 
+
 class FlutterRustorePushService : RuStoreMessagingService() {
     companion object {
-        var client: FlutterRustorePushClient? = null
+        var client: RuStorePushCallbacks? = null
     }
+
+    private val uiThreadHandler: Handler = Handler(Looper.getMainLooper())
 
     override fun onNewToken(token: String) {
         Log.d("onNewToken: $token")
 
-        client?.onNewTokenResult?.success(token)
+        uiThreadHandler.post {
+            client?.newToken(token) { }
+        }
     }
 
     override fun onMessageReceived(message: RemoteMessage) {
         Log.d("onMessageReceived: $message")
 
-        val result = Rustore.Message.Builder()
-        result.setMessageId(message.messageId)
-        result.setData(message.data)
-        result.setPriority(message.priority.toLong())
-        result.setTtl(message.ttl.toLong())
-        result.setCollapseKey(message.collapseKey)
-        result.setNotification(
-            Rustore.Notification.Builder()
-                .setTitle(message.notification?.title)
-                .setBody(message.notification?.body)
-                .setChannelId(message.notification?.channelId)
-                .setClickAction(message.notification?.clickAction)
-                .setIcon(message.notification?.icon)
-                .setColor(message.notification?.color)
-                .setImageUrl(message.notification?.imageUrl.toString())
-                .build()
+        var notification: Notification? = null
+
+        if (message.notification != null) {
+            notification = Notification(
+                title = message.notification?.title.orEmpty(),
+                body = message.notification?.body.orEmpty(),
+                channelId = message.notification?.channelId.orEmpty(),
+                clickAction = message.notification?.clickAction.orEmpty(),
+                icon = message.notification?.icon.orEmpty(),
+                color = message.notification?.color.orEmpty(),
+                imageUrl = message.notification?.imageUrl.toString(),
+            )
+        }
+
+        val msg = Message(
+            messageId = message.messageId,
+            data = message.data as Map<String?, String?>,
+            priority = message.priority.toLong(),
+            ttl = message.ttl.toLong(),
+            collapseKey = message.collapseKey,
+            notification = notification
         )
 
-        client?.onMessageReceivedResult?.success(result.build())
+        uiThreadHandler.post {
+            client?.messageReceived(msg) { }
+        }
     }
 
     override fun onDeletedMessages() {
         Log.d("onDeletedMessages")
 
-        client?.onDeletedMessagesResult?.success(null)
+        uiThreadHandler.post {
+            client?.deletedMessages { }
+        }
+
     }
 
     override fun onError(errors: List<RuStorePushClientException>) {
         Log.d("onError: $errors")
 
-        client?.onErrorResult?.success(errors.toString())
+        uiThreadHandler.post {
+            client?.error(errors.toString()) { }
+        }
     }
 }